MacOSX10.14.sdk muduo c++11 编译base net 库成功。

+++ muduo/base/Condition.cc
@@ -9,7 +9,7 @@
 #include <sys/time.h>
 
 // returns true if time out, false otherwise.
-bool muduo::Condition::waitForSeconds(int seconds)
+bool muduo::Condition::waitForSeconds(double seconds)
 {
   struct timespec abstime;
 #ifdef CLOCK_REALTIME
